test: add initial testnet module
This is currently intended to provide a minimal application on top of
comet, to observe what happens when trying to run multiple comet
instances in a single process.

The test added in this commit is likely more complicated than it needs
to be, but the test is finally running so it is an appropriate
checkpoint.